[feat] : clean the code

+++ kernel/defs.h
@@ -174,7 +174,7 @@ int             copyout(pagetable_t, uint64, char *, uint64);
 int             copyin(pagetable_t, char *, uint64, uint64);
 int             copyinstr(pagetable_t, char *, uint64, uint64);
 int             is_cow_page(pagetable_t pagetable, uint64 va);
-int             alloc_cow_page(pagetable_t pagetable, uint64 va);
+uint64          alloc_cow_page(pagetable_t pagetable, uint64 va);
 // plic.c
 void            plicinit(void);
 void            plicinithart(void);
